+++ lib/extensions/widget_ext.dart
@@ -315,6 +315,71 @@ extension WidgetExtensions on Widget {
 
   // ? ================ Shapes & Boxes ================
 
+  /// Returns a new Scaffold widget with the given body.
+  Scaffold scaffold({
+    Key? key,
+    PreferredSizeWidget? appBar,
+    Widget? body,
+    Widget? floatingActionButton,
+    FloatingActionButtonLocation? floatingActionButtonLocation,
+    FloatingActionButtonAnimator? floatingActionButtonAnimator,
+    List<Widget>? persistentFooterButtons,
+    AlignmentDirectional persistentFooterAlignment =
+        AlignmentDirectional.centerEnd,
+    BoxDecoration? persistentFooterDecoration,
+    Widget? drawer,
+    void Function(bool)? onDrawerChanged,
+    Widget? endDrawer,
+    void Function(bool)? onEndDrawerChanged,
+    Widget? bottomNavigationBar,
+    Widget? bottomSheet,
+    Color? backgroundColor,
+    bool? resizeToAvoidBottomInset,
+    bool primary = true,
+    DragStartBehavior drawerDragStartBehavior = DragStartBehavior.start,
+    bool extendBody = false,
+    bool drawerBarrierDismissible = true,
+    bool extendBodyBehindAppBar = false,
+    Color? drawerScrimColor,
+    Widget? Function(BuildContext, Animation<double>) bottomSheetScrimBuilder =
+        _defaultBottomSheetScrimBuilder,
+    double? drawerEdgeDragWidth,
+    bool drawerEnableOpenDragGesture = true,
+    bool endDrawerEnableOpenDragGesture = true,
+    String? restorationId,
+  }) {
+    return Scaffold(
+      appBar: appBar,
+      backgroundColor: backgroundColor,
+      bottomNavigationBar: bottomNavigationBar,
+      bottomSheet: bottomSheet,
+      bottomSheetScrimBuilder: bottomSheetScrimBuilder,
+      drawer: drawer,
+      drawerBarrierDismissible: drawerBarrierDismissible,
+      drawerDragStartBehavior: drawerDragStartBehavior,
+      drawerEdgeDragWidth: drawerEdgeDragWidth,
+      drawerEnableOpenDragGesture: drawerEnableOpenDragGesture,
+      drawerScrimColor: drawerScrimColor,
+      endDrawer: endDrawer,
+      endDrawerEnableOpenDragGesture: endDrawerEnableOpenDragGesture,
+      extendBody: extendBody,
+      extendBodyBehindAppBar: extendBodyBehindAppBar,
+      floatingActionButton: floatingActionButton,
+      floatingActionButtonAnimator: floatingActionButtonAnimator,
+      floatingActionButtonLocation: floatingActionButtonLocation,
+      key: key,
+      onDrawerChanged: onDrawerChanged,
+      onEndDrawerChanged: onEndDrawerChanged,
+      persistentFooterAlignment: persistentFooterAlignment,
+      persistentFooterButtons: persistentFooterButtons,
+      persistentFooterDecoration: persistentFooterDecoration,
+      primary: primary,
+      resizeToAvoidBottomInset: resizeToAvoidBottomInset,
+      restorationId: restorationId,
+      body: this,
+    );
+  }
+
   /// Returns a new constrained widget with the given width and height.
   SizedBox constrained({double? width, double? height}) {
     return SizedBox(width: width, height: height, child: this);
@@ -589,4 +654,11 @@ extension WidgetExtensions on Widget {
       child: this,
     );
   }
+}
+
+Widget? _defaultBottomSheetScrimBuilder(
+  BuildContext context,
+  Animation<double> animation,
+) {
+  return null;
 }